extract XmlErrorException
